Moved all formatting to formatter
Formatter object now deals with all formatting logic
