+++ src/main/groovy/org/mockito/release/notes/GitNotesBuilder.java
@@ -39,7 +39,6 @@ class GitNotesBuilder implements NotesBuilder {
         ImprovementsProvider improvementsProvider = Improvements.getGitHubProvider(authTokenEnvVar);
         Collection<Improvement> improvements = improvementsProvider.getImprovements(contributions, labels);
 
-        return new NotesPrinter().printNotes(version, new Date(),
-                DefaultFormatter.format(contributions), DefaultFormatter.format(labels, improvements));
+        return DefaultFormatter.formatNotes(version, new Date(), contributions, labels, improvements);
     }
 }
